Using INT

+++ IFES-4-PERIODO/ESTRUTURA DE DADOS/TRABALHO/pilhas.c
@@ -5,10 +5,10 @@
 
 struct Pilha {
     int topo;
-    float itens[TAMANHO_MAX];
+    int itens[TAMANHO_MAX];
 };
 
-void empilhar(struct Pilha *p, float valor) {
+void empilhar(struct Pilha *p, int valor) {
     if (p->topo < TAMANHO_MAX) {
         p->itens[p->topo] = valor;
         p->topo++;
@@ -17,8 +17,8 @@ void empilhar(struct Pilha *p, float valor) {
     }
 }
 
-float desempilhar(struct Pilha *p) {
-    float valor = 0.0;
+int desempilhar(struct Pilha *p) {
+    int valor = 0;
     if (p->topo > 0) {
         p->topo--;
         valor = p->itens[p->topo];
@@ -28,8 +28,8 @@ float desempilhar(struct Pilha *p) {
     return valor;
 }
 
-float retornatopo(struct Pilha *p) {
-    float valor = 0.0;
+int retornatopo(struct Pilha *p) {
+    int valor = 0;
     if (p->topo > 0) {
         valor = p->itens[p->topo - 1];
     } else {
@@ -42,7 +42,7 @@ int main() {
     struct Pilha minhapilha;
     minhapilha.topo = 0;
     int op;
-    float valor;
+    int valor;
 
     while (1) {
         printf("\n1 - Empilhar(push)\n");
@@ -55,25 +55,25 @@ int main() {
         switch (op) {
             case 1: // push
                 printf("\nDigite o valor a ser empilhado: ");
-                scanf("%f", &valor);
+                scanf("%d", &valor);
                 empilhar(&minhapilha, valor);
                 break;
 
             case 2: // pop
                 valor = desempilhar(&minhapilha);
-                if (valor != 0.0) {
-                    printf("\n%1f DESEMPILHADO!\n", valor);
+                if (valor != 0) {
+                    printf("\n%d DESEMPILHADO!\n", valor);
                 }
                 break;
 
             case 3: // mostrar topo
                 valor = retornatopo(&minhapilha);
-                if (valor != 0.0) {
-                    printf("\nTOPO: %1f\n", valor);
+                if (valor != 0) {
+                    printf("\nTOPO: %d\n", valor);
                 }
                 break;
 
-            case 4: // sair
+            case 4:
                 exit(0);
 
             default:
